requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
streamlit>=1.0.0
numpy>=1.21.0
scikit-learn>=0.24.2
